Package image component and add testing framework
