+++ packages/hydrogen/package.json
@@ -13,7 +13,8 @@
   "scripts": {
     "build": "tsup --clean --config ./tsup.config.ts",
     "dev": "tsup --watch --config ./tsup.config.ts",
-    "typecheck": "tsc --noEmit"
+    "typecheck": "tsc --noEmit",
+    "test": "jest"
   },
   "exports": {
     ".": {
@@ -54,6 +55,12 @@
   },
   "devDependencies": {
     "@shopify/hydrogen-react": "*",
-    "@types/recursive-readdir": "^2.2.1"
+    "@types/recursive-readdir": "^2.2.1",
+    "jest": "^29.3.1",
+    "@jest/globals": "^29.3.1",
+    "babel-jest": "^29.3.1",
+    "@babel/core": "^7.20.12",
+    "@babel/preset-env": "^7.20.2",
+    "@babel/preset-typescript": "^7.18.6"
   }
 }
