+++ packages/hydrogen/src/index.ts
@@ -10,3 +10,4 @@ export {InMemoryCache} from './cache/in-memory';
 
 export {RESOURCE_TYPES, REQUIRED_RESOURCES} from './routing/types';
 export {notFoundMaybeRedirect} from './routing/redirect';
+export {Image} from './image';
